Re-write _onData to handle messages and raw data split over packets
Note - the reading of the RTSP message has to be one byte at a time
as there is no length information.
However the reading of the RTSP Payload (the SDP data) and the
interleaved RTP and RTCP packets can be changed to use array copies as the
payload size is known.