The Session should be left as String.
Session can have an optional ';timeout=nn' part so string operations such as split(';')
are performed on the Session string. If the session was all numbers, yellowstone would
convert it to an integer and the string operations will fail.
This also fixes a potential issue where the Session could be 00001234.
Yellowstone would convert it to an integer of 1234 and then try and access Session 1234
instead of Session 00001234.