Fix absolute media stream URLs
Needed to make an Axis V5915 IP camera work
Note in the SDP is a 'control' field.
If it is relative (eg control: streamID=01) then you add it
to the end of the RTSP URL to get the media stream URL.
But some times the control field is a full RTSP url. So you
use it directly