Fix Digest Authentication Error.
Thanks to @Leone25 for the bug report.

A previous commit made Yellowstone only handle RTS Authentication correctly for the
first 2 RTSP Commands (test of id > 2).
This fixes that change so that we only reject() the promise if we had included
an Authorize Header in the outgoing RTSP command and we get a 401.

Note there is an optimisation here that @	Leone25 spotted.
We could cache the Digest Realm and Nonce and then insert the Authorize
Header at the start of each RTSP Command
to save the double round trip of sending the
RTSP Command, getting a 401 and sending it again.
The Authorizaition Header for BASIC can be cached.
For Digest it needs to be re-computed as it includes the RTSP Command in
the Password Hash